Fix set_field
get_hex returns an immutable reference, and since Hex is an array and arrays
implement Copy, set_field was modifying a copy of a hex instead of the actual
board. It doesn't seem right to make get_hex return a mutable reference, since
other functions just want to inspect the hex and not modify it. So, for now,
set_field is fixed by making it modify the copy of the hex and then using
set_hex to update the board.