Replace .is_none checking with insert_if_empty
It's a bit annoying to have to pass &mut event, but the only other method on
Option<T> which might do the trick is get_or_insert. But, that takes another
value of type T to insert, not an Option<T>. So, it add extra .is_some checks.

I also tried to use a closure that would borrow &mut event inside its body, so
that we wouldn't have to pass in the reference each time. However, because of
non-lexical lifetimes (I think that's the problem), the reference is captured by
the closure for the whole scope of the function. So, we're unable to return
event at the end. The only way to fix this is to wrap the closure and all the
places it's used in one big block--excluding the definition and returning of
event. Then, the reference is dropped before we want to return event. However,
that's even more ugly, so I've just opted for passing in the reference each time
to a helper function.