Implement undo and redo
can_remove_piece was added back, and remove_piece/exchange_piece no longer
return bools, because model.commit_move needs to be called before any moves are
done. Board can't call methods on Model, so update needs to know if a move is
going to happen before it actually executes it.

The logic for ensuring that invalid Events aren't generated (e.g. Clicks aren't
generated when the game is over) has been shifted to view::draw. This is
slightly dangerous, as it's possible for invalid Events to reach update.
However, this also makes logic in update cleaner, as multiple match blocks with
conditional behavior on GameResult aren't needed.

The logic controlling button placement in draw continues to grow uglier. Without
real disableable buttons (faking it via alpha doesn't work, because the button
will still change alpha upon mouse over and click--just while being paler), the
cleanest option is to hide buttons that can't be used. Since there is no
BeginGroup/EndGroup counterpart to same_line, nested ifs are used to ensure that
the undo/redo buttons are always on the same line, even when one or the other is
missing. This will eventually be fixed when disableable buttons are added, or a
better hack is found.