Rewrite is_hex_removable
The previous attempt to rewrite is_hex_removable to be more elegant
(a816eaf) introduced a bug: it defined a removable hex as having no
solitary neighbors, i.e. no neighbors who were not adjacent to at least
one other neighbor.

This definition overlooks the fact that hexes with only one neighbor are
removable, even though all of their (single) neighbors are solitary.

The new implementation fixes this problem. While it isn't elegant, it is
simple to understand and verify, and still efficient. So, it is an
overall improvement.