Integrate AI via threads
The AI runs on a separate thread so that the view stays responsive while the AI
is thinking. However, because Rust doesn't provide a platform-independent way to
kill a thread, a mpsc::channel is used to send a stop message to the AI thread.

Now, I didn't test this, but my guess is that calling try_recv on every
iteration of negamax would be slow. So, the AI only checks to see if it has been
stopped in the top ai_move function--that is, once per root move examined. This
means that the AI may not terminate immediately, depending on where it is in the
search.

Usually, this isn't a problem, as the AI runs fairly quickly. However, assume
an extreme case where the AI took a few minutes to run, and the user spammed the
New Game "Computer vs. Human" option. Many AI threads might be running
simultaneously, and slow down each other.

Thus, when ai_move is called, one of the arguments passed in is the handle from
the previous AI thread. If this handle is Some, the new AI thread will stop the
old thread and wait for it to finish before running its own search. While this
does slow the new AI thread, it ensures that there is only ever one AI thread
running at once. This prevents the AI threads from taking up too many resources.