Add Outcome, block invalid Events in update
Games won by resignation or an offered and accepted draw probably should not
change board.outcome. After all, they are outcomes caused by external events,
and are inconsistent with the actual state of the board. Still, the alternative
is worse: have Outcome in Board and a similar enum in Model to track such
resigns and draws. That would be confusing. So for now, board.outcome will
always contain the outcome of the *game*, and not necessarily the board.

Previously in 70a5403, the logic for blocking invalid Events was moved to draw.
Now it has been moved back, for two reasons. First, draw shouldn't concern
itself with what Events are valid in what situations, except for the cases like
the exchange button and available moves, where the UI elements drawn must
directly correspond to the valid moves a player can make. Second, if the invalid
Event checking logic isn't in update, it isn't clear by looking at update that
updates doesn't have a bug. Also, it's just safer and simpler to validate input
instead of trusting other functions to produce valid input.