Change AI score type to i16
This likely has no performance benefits. However, it does make the constants
such as INFINITY easier to reason about. Also, there is no danger of going
out-of-bounds. If a piece is 100 points, a player would have to be winning by
about 164 pieces to exceed the LOSE constant, and about 328 to overflow the i16.