Rename PlayerType to Player
This name is a bit ambiguous, as people often call "White" and "Black" the
players. However, we use Color for that, and since I can't think of a word that
encompasses humans and computers, I'll stick to Player. Adding "Type" was my
attempt to make this distinction between colors and organic/silicon agents, but
it doesn't work: how could an enum not be a type?