Reorder expression in is_valid_coord
By resizing the board to an absurdly small height and a large width, it was
possible to hover over a pixel that converted a large coordinate like (90, 90).
is_valid_coord would then panic when trying to add the coordinates together.
This puts the other checks first so that (x + y).abs() will no longer panic.