Block on the TTable mutex
If an AI thread has been sent the stop signal but has not received it yet, it
will still hold on to the TTable mutex. The next AI thread to start will not
join on the old thread's handle because AI::stop throws away the old Status
enum. The new thread will then try_lock and panic because the old thread is
blocking the acquisition.

Now, any new thread will block on locking the TTable mutex. This makes storing
the handle a bit redundant, but it's nice to do so anyways, as it allows for one
to wait for the thread to finish without having to acquire the TTable at the
same time (should one want to do that).