Replace Board getters/setters with pub
Model wasn't using getters/setters, so why should Board? This makes the code
simpler and shorter (and maybe even slightly faster). The only downside is that
we have to trust that other code won't mess with the Board members--but that
seems reasonable give that this is a game engine (knives vs files or something
like that).