Don't check TTable in quiescence search
This caused massive slowdowns because the replacement strategy would (probably)
never commit any quiescence entries due to their depth of 0. This meant that
every TTable probe was a pointless cache miss. It might be possible to encourage
the TTable to keep quiescence entries around longer through a better replacement
scheme. But not storing them at all is much simpler.