Store hex capture moves in BitBoard, not array
It seems that the map with the closure that captures hex_capture_moves was quite
slow. My guess is that it was moving the whole array into the closure? This
method requires an extra HEX_MASK lookup, but copies a BitBoard instead of an
array. It also doesn't require a huge comment to explain.