Remove iterator madness TODO comment
This iterator is still pretty inefficient (it seems that nested iterators in
general, like flat_map and chain, ahem, are not that efficient). However, it has
gotten much better after moving much of the work to out of the iterator. Though
it's still a hot spot, this reminder is no longer needed.