Reorder generate_capture's iterator chain
The exchange_captures are always valid moves, unlike the other two iterators,
which may fail if there are no pieces to capture or empty fields to move to. By
placing the guaranteed moves first, we might be able to achieve cutoffs in the
quiescence search that allow us to skip the other two expensive iterators.