Reorder REMOVABLE_HEX_COMBS for corner/edge hexes
Corner and edge pieces have only 6 and 9 combinations to test for, respectively,
and not 18, like center pieces. This reordering places those cases at the front
so that there are fewer unnecessary checks.

The benchmarks are disappointing, though, because while this is basically the
optimal ordering for this system, the program only runs about 1.01x faster.