Fix capture generation
This fixes a big bug that generated lots of false captures, which *completely*
bogged down quiescence search. As an example, in one of the (many) slow
positions (45s to search 6 plies), quiescence nodes represented 88.8% of nodes
searched! Even worse, the quiescent nodes at depth 5 represented 99.2% of all
quiescent nodes! (88.1% of all nodes searched.)

With this change, in that example position, the number of nodes searched falls
to 53% of the previous number. The search is still mostly quiescent nodes
though: 83.9% of nodes are quiescent, and 98.9% of those come from depth 5.