Use TTable in quiescence, new replacement strategy
To speed up quiescence search, it now consults the transposition table. The
depth of the root quiescence node is 0, and each ply under it is -1, -2, etc.
The replacement strategy is to now replace if the age is different (which means
to replace every turn) or if the depth of the replacement is greater. This
allows quiescence entries to actually get into the table, and means that
quiescent nodes can actually use the scores from non-quiescent nodes.