Check stop signal before sending move
If the AI is sent the stop signal after already beginning to search the last
root move, it will not receive it and try to send the move anyways. This will
cause the thread to panic because the sender will be disconnected. This will
then poison the TTable mutex, which will cause the next AI thread to stop and
basically ruin everything.