Only draw UI on mouse event, remove unused code
We now use run_forever instead of poll_events to only redraw the UI when
something happens (this works because this game has no animations). For cursor
movements, we cap the framerate at about 25fps to save CPU time. For mouse
clicks we draw the UI twice to immediately show the results of the click. We
also draw one frame before the event loop begins so that the screen isn't empty.

Also, all the keyboard code has been removed, since it's unlikely this game will
ever support the keyboard or require a text entry.