Redraw on AI move
Previously, the game was very efficient, but when playing against the AI, it
wouldn't update unless you continuously moved the mouse around. This passes an
EventLoopProxy to the AI thread so that it can wake up the main thread when it
has finished thinking. As usual, we need to render calls--one to update the
Model and one to show it.