Show AI debug info in window, use RefCell in Model
imgui likes being passed mutable references. However, view is given &Model. So
we use RefCells to mutate the search depth and "show debug info" boolean.

The debug info string is a RwLock even though there's only one reader and one
writer. It just seemed to be the right struct to use, anyhow. The string is also
cloned each frame because it seems there's no other way to give it to imgui.

The mouse scroll code also had to be added back in so that the debug info window
could be scrolled.