Rename some instances of "hex" to "tile"
It turns out that what I've been calling a "hex" for the entire duration of this
project is actually, in official wording, a "tile". For the sake of clarity, a
six-sided regular polygon will always be called a "tile" in the README or game
UI. However, I will not change the name internally from "hex" as it's too much
effort.