Upgrade to imgui v0.1.0
Not sure why everything but the piece border was unaffected by the DPI
stuff that the windowing crate is now doing. Doubling it seems to fix
things but I'm not sure if that's only going to work on my system. Also
not sure if touch events will continue to work. I adjusted them by DPI
because it seemed like the right thing to do, but who knows.
In conclusion, the view::sys code grows more and more convoluted and I'm
less and less sure of what it does.