Prevent coordinate check overflow
If you resize the window so that the board is very small, attempting to
convert the mouse position to a HexCoord results in very large values.
Hitting exactly -128 in either coordinate causes a panic in debug mode
when .abs() is called. So, we use RangeInclusive instead.
Now, maybe we shouldn't care about this as it's just a debug panic. But
it turns out the coordinates (-128, -128) will actually pass the checks
in release mode:
  * (-128i8).abs() == -128 <= 2
  * (-128i8 + -128i8).abs() == (0i8).abs() == 0 <= 2
So, it's important that we prevent this from happening.