Enforce minimum board size
The board doesn't look good at ridiculously tiny sizes. In addition to
being too small to see or play, the piece border thickness and hex gap
equations break down. And, making the window very small causes the board
to flip upside down, which messes with the appearance of the pieces.